Setup Starlight CI
CI builds are currently broken down into three stages:

  - `prebuild`: a manual run only stage to rebuild the surfliner `builder`
    image. This image is specified in the base `./Dockerfile` and is setup to
    run `docker` and `docker-compose`, for building other images.
  - `build`: Holds each product's build stage (if any).
  - `test`: Holds the actual test suite runs.

We may want additional stages later, e.g. `lint`.

The initial setup is for `starlight` only. The builds trigger when the
`./starlight` directory (or `./Dockerfile` or `.gitlab.yml`) changes. Builds for
other projects can be set with similar filters.

The actual build currently fails, due to gaps in the Solr configuration. We
should resolve this against !7 (or at least agree on a path to resolution)
before merge.