[lark] Bootstrap a Rack application
Adds a basic `Lark` application and an initial route. The route `/:id` currently
has a hard coded response of `404` with empty body and headers. This passes an
initial test, but we still need a test case for dummy data and a non-404
response structure.

We're starting with a simple Rack app (using Sinatra) so we can see how complex
things get before falling back on Rails.

Connected to #25.

Committed with @soops, @tchu.