lark: Add a `Concept` model
The `Concept` model has a single attribute (`pref_label`).

We wire the model in through the `ResourceController` to allow reasonable
response bodies. Currently, the response body is simply the `#to_json` on the
concept. We'll need to refine this with some kind of view code as a next step.

Committed with @tchu, @lsitu.