lark: Add pub/sub to the EventStream
Uses `dry-events` (this name runs the risk of being confusing) to support
internal pub/sub for events. Listeners can be added to the stream to be notified
when events are appended to it. The initial use case for this is to support
indexing: an "index listener" can be added to watch for events that require
changes to the index.

We could also use this interface to notify users of changes, enqueue background
jobs, etc... when appropriate events are added to the stream.

A factory is added for the `:Create` event type, to support testing.

Related to #28.