lark: Add id minting with UUID
When doing a `#create` via POST, mint an ID and return it in a json
response. The ID is (temporarily) a UUID.

`dry-transaction` is introduced to handle the 'business transaction' called by
the controller. This includes minting the ID, logging the `:create` event, and
building an object for the response.

Connected to #26.