lark: Add error handling and HTTP response for unknown request bodies
Add an error handling approach for code below the
controller(s). `Lark::RequestError` provides a `#status` which can be used to
determine the correct HTTP response codes for errors raised throughout the
application.

A subclass (`UnsupportedMediaType`) is raised when a non-JSON POST body (or one
without a parser) is sent. This leads to a 415 response code; see:
https://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html.

Other subclasses can be added (and raised) to emit other HTTP response codes.

Tests are added to the API, Controller, and `RecordParser#for` specs.

Connected to #26.

Committed with @soops.