lark: Refactor response serialization to use a `RecordSerializer`
A `RecordSerializer` is introduced for building response bodies from `Concept`
and other authority record models. The idea is that, given the correct options,
a serializer and a parser should be able to round trip records (i.e. a record
serailized and then parsed should yield an identical `Concept`
object).

For now, the implementation is hard-coded. It's a straight refactor of the
behavior previously in `RecordController#show` and `#create`.