lark: Add real-server integration tests
Setup a test suite for integration tests hitting a real server over HTTP.

We use `capybara_discoball` which makes it easy for us to stand up a server in
the context of our test suite. Tests then hit the server with requests using
`Net::HTTP` (included in Ruby core).

Our test infrastructure is arranged to avoid starting the server repeatedly when
more tests are added.

The initial spec represents work we demoed at the end of __Sprint 1__.

Fixes #60.

Committed with @tchu1.