lark: Refactor update authority tx to set adapter on initalize
This is an attempt to tighten up the `UpdateAuthority` transaction interface by
reducing the amount of data that needs to be passed in the
`Result` (`Success`/`Failure`) monad. Since we'll be passing the adapter through
the entire transaction anyway, it seems better to set it once, up front, and
avoid passing it through step `#call`s.

The new approach is also in line with how `event_stream` is handled.