lark: Rename and document `:update` event to `:change_properties`
Proposes a rename of `:update` to `:change_properties`, and adds a required
`changes` key to the event's data. Initial documentation is added for the new
event type.

This is the outcome of discussion with @lsitu; but the proposed new name is
somewhat arbitrary. Suggestions are welcome. It's important that event
semantics get nailed down before any production events go in the log. Once
they are in, we need to support them for the life of the app.