starlight: Move `dotenv-rails` to `:development` / `:test` group
[`dotenv`](https://github.com/bkeepers/dotenv#dotenv---) is great for setting
environment variables in _development_, but we should be less opinionated about
how we load them in production. We want adopters to be able to set environment
variables using local deployment processes and secrets infrastructure.

This gem was initially introduced by DCE, and presumably was used in their
deployment processes. This is a step towards us being able to use it as well.

Connected to #64.