lark: Add custom `:create` strategy for FactoryBot for Valkyrie
Because `valkyrie` doesn't use an on-object `#save` method, the default
`FactoryBot` create strategy doesn't work for it. A custom strategy that uses
the app-wide index adapter is added here to work around this.

We register this as `:create` to give transparent use with the existing
`FactoryBot.create(object)` syntax.

Committed with @lsitu.