starlight: Set environment variables for CI's Solr image
Sets up environment variables for retrieving the Solr config from the
repository. `CI_COMMIT_SHA` and `CI_REPOSITORY_URL` are set in the GitLab
Runner's build environment, but not in the service containers
themselves. Assinging their values to variables within the containers allows
access to them there.

This uses the base Solr image configured by !63. Since it's built and in the
registry already, there is no need to duplicate the relevant changes
here (supposing, of course, that !63 is eventually merged).

See also !63. Connected to #57.