[lark] Add an initial Helm chart for lark
Introduces a Helm catalog at `/charts` with a custom chart (see:
https://rancher.com/docs/rancher/v2.x/en/catalog/custom/) for deploying
`lark`.

The application runs using the image from
`registry.gitlab.com/surfliner/surfliner/lark_web:stable`, which is built
whenever we merge to master with changes to `lark/**/*` (this may need
refinement). For now, this is the only container that runs as part of the
deployment.

This chart uses the Rancher `questions.yml` to support launch options. For now,
the only "question" is a chooser for index backends, but we'll want to extend
this shortly to select a hostname for the ingress.

Related to #34.