lark: Setup basic review and staging deploys
Use GitLab auto-DevOps-like functionality to deploy lark.

Adds a `review` stage that supports short-lived deployments for each branch/tag
pushed, as well as a `staging` deployment that is triggered on merge to `master`.

The basic behavior we reuse is at
https://gitlab.com/gitlab-org/gitlab-ce/blob/master/lib/gitlab/ci/templates/Jobs/Deploy.gitlab-ci.yml#L59,
with some changes to meet our internal needs:

  - We retain our rebuild of `stable`. This could use refinement as we work on
    our review deployments.
  - We don't install dependencies at the beginning of the deploy process. Our
    base `builder` image has the dependencies installed already.
  - Since our charts are maintained in the monorepo we skip downloading the
    chart, simply copying it to the location GitLab's auto-deploy expects.
  - Our URL is temporarily `lark-staging.local`. This also needs work.

More work is needed to setup the Solr and Postgresql variables/questions on deploy.