lark: Make staging deploy required; use `master` tag
The documented approach to image tags is that we use `master` for the current
`master` deployment, and `stable` for production. This fixes the implemented tag
approach to match that approach, and deploy `master` to the staging environment.

The solr and postgres values ('answers') are also fixed to repair the staging
deployment, and that deployment is made required for a "successful" master
build.

This completes work on #131.