lark: ADd `:database` application config via `Dry::Configurable`
We're using `Dry::Configurable` to store other application configuration at
runtime. This consolidates by adding `database` to that config. This
configuration value is left empty unless the `:sql` adapter is in use.

We retain the existing `config/database.rb` to allow the rake tasks to create
the configuration without loading the entire application.