lark: remove defaults for database configuration
Using `ENV.fetch` ensures that missing config eagerly fails, instead of using
undesirable defaults. For the rake tasks, this means we should decline to load
the database tasks if the `:sql` adapter is not in use. This way, an environment
still doesn't need to provide settings that won't be used. A guard is added to
achive this.