[lark] Make `scheme` an internal attribute
We need default value behavior for the `:scheme` attribute. The old strategy
won't work in `Dry::Struct` 1.0 (and probably should never have worked).

Since we'll want to better define the scheme behavior and its relationship with
the `Valkyrie::Resource` classes later on, we temporarily remove `scheme` from
the configured metadata and treat it as an internal value.