surfliner: Proposed style guide changes for Rails projects
Rails 5.2.x uses top-level includes in `bin/setup` and `bin/update`; this allows
that explictly. In general, it should be okay to do this in `bin/*` scripts.

Also excludes the contents of `config` from blocklength rules. Config files
often use a configuration DSL, which results in longer blocks that would be used
in general purpose code.