surfliner: remove code-quality CI jobs
The current code-quality jobs are using a container image that
ultimately runs a Code Climate docker image.

This Code Climate image is, ultimately, doing the same job that our lint
jobs are already doing, namely running rubocop. While there are
theoretically more features available in Code Climate, we are not
electing to use them, and therefore these jobs are redundant.