shoreline: don't persist download derivatives across deploys/pods
Using a `ReadWriteOnce` volume is preventing 0-downtime deploys (and, less
importantly, multi-pod Rails installs). Once the volume is attached, we can't
deploy another pod that will depend on it. Ideally, we'd be able to share the
volume safely across Ruby/Rails pods, but we're punting that requirement for the
short term in favor of a simple `emptyDir` for the downloads cache.