surfliner: refactor and reorganize CI/CD config
further break-out the ci/cd configuration by project. this should reduce the
number of future edits to the top-level pipeline configuration, avoiding
unneeded job runs. with this organization, we should only need to edit the
top-level config when we are adding a new project or want to make changes to
shared scripts (we could consider moving these and requiring them per-file in
the future, too) .