shoreline: tweak geoserver liveness/readiness probes
these probes seem a bit unrealiable still!

possibly the cause is a long response time on initial requests? the default
kubernetes timeout is 1 second, bump that to 5 for maybe a more tolerant
startup?

we probably don't want to consider this service healthy when it's taking 5
seconds to respond, but let's not shut it down so eagerly. as long as we only
have one node, let's be super conservative about when to give up on it.