surliner: [style] allow constant definition in RSpec files
RuboCop wants us to avoid constant definition in RSpec files, using class/module
stubs instead. This isn't really sufficient for some pretty basic
interaction/integration test use cases (or, rather, it becomes wildly complex
when we need a real class with a real name, etc...; especially in Rails).

In general, a good pattern for defining constants is to clean them up with
`remove_const`. To avoid mucking about with the Kernel/Object namespaces, and
keep tests from stepping on each others' toes, a good pattern is to define a
module within `MyProject::Test` to act as a namespace scope for the file, then
send `.remove_const` to strip the whole namespace. this can be done in an
`after` block. The pattern is documented in the rubocop config file to clarify
intent.