starlight: remove file-based csv upload feature
Since we are deploying Starlight in k8s it no longer makes sense to
support this BINARY_ROOT/file-based csv upload feature in the codebase.
We had already stopped using it with ansible-based deploys, but now it
really makes sense to remove this to avoid any confusion in features for
other interested parties.