shoreline: use shapefile name as layer id
GeoServer will set the layer name to the name of the .shp file that is uploaded
by our rake task.  Until now, it happened to be the case that the names of the
zipfiles were the same as the names of their contained .shp files, but that is
no longer the case.  As a result, the layer ID being set in GeoBlacklight
differed from the actual layer name, and GeoBlacklight was unable to fetch
shapefiles from GeoServer.