starlight: make sidekiq containers wait for pending migrations
in the past, the sidekiq deployment has been configured to wait for the webapp
deployment to be ready. the reasoning for this is that the webapp deployment
runs the database migrations, which are needed to start sidekiq.

this refactor makes this wait explict. instead of waiting for the web service,
this now checks for db availability and migration status directly. this should
be more resilent as db setup becomes decoupled from the web service.