surfliner: don't depend on `git` in gemspecs
we don't normally include .git files in containers (to save on size, avoid
internal complexity in running systems, etc...). since we need to build gems in
containers, it's important we not rely on git as part of the build process.

to avoid this, we specify the files we want to include in the gem more
declaratively. this should also help us avoid packaging unintended files in
gems.