starlight: bump universalviewer@3.0.22 to universalviewer@3.0.23
universalviewer@3.0.22 ignores the `config` option in the URL, so it
didn't matter that app/views/catalog/_universal_viewer.html.erb
hardcoded a config file that does not exist. Starting in
universalviewer@3.0.23, the missing config file results in errors. This
change updates universalviewer to 3.0.23 and removes the hardcoded
config file URL parameter.