push knowledge about HTTP level errors into the controller
hides some knowledge about HTTP errors and status codes from the
`CreateAuthority` transaction.

instead of returning `BadRequest` directly (this gives the wrong status code!),
returns a `Failure` result from the transaction and handles the error case
within the controller. other steps within this transaction still know about
`BadRequest`, and we may want to follow up by doing a similar refactor for
them.

also takes a hack at standardizing `Failure` result data, in the form of a
custom RSpec matcher. that should be useful in the follow up.

as support, introduces a `FailureMinter` so we can aggressively fail minting at
will. ~*we are so powerful!*~