lark: move custom rspec matcher to `spec/support` and add docs
recent work added a custom RSpec matcher. it would be good for other tests
to reuse that matcher, so we wanted to move it to a shared code location (it had
been inline with the tests that use it). since we already use `spec/support` for
`FactoryBot` build strategies, `spec/support/matchers` seemed like a good home.

the custom matcher is one half of a `Dry::Monads[:result]` test setup. putting
the actual test in `result.rb` should make it easier to load all related support
code in the future, assuming we end up writing support code for `Success`
results.

docs are added to clarify intended usage and point folks to the handy, built-in,
`be_success` and `be_failure` matchers.