lark: test stream failure behavior, as well as invalid attributes
new validation behavior made it in to fix #671, but this change made the test
for `KeyError` incorrect. we want to ensure this transaction fails for both the
case where the stream errors, and in the case of validation failures.

this rewrites the original test to make the correct ensurances, and adds a new
test for the validation behavior.