lark: support symbol based resolution for `Authority` schema
we've started to see some alignment between `AuthorityContract` and the
`Authority` models in terms of the way they translate YAML configs to
`dry-schema`.  it's not totally clear where all the alignment will lay, but we
want to start to extract code out so we can eventually support M3 in both of
these classes.

this is a first step, just getting some code duplication in place so we can
extract it.