lark: use pg admin user to create db/migrate
Due to the requirement of installing the PG uuid extension, the
`db.rake` file needs to run it's tasks as the admin `postgres` user
using elevated permissions.

So, we're introducing a DATABASE_AS_ADMIN hash which holds this
information and can be passed to Sequel when creating the database,
installing the extension, and running the migrations. The `postgres`
user then `GRANT`s access to our regular postgres user on the various
tables so that user can be used by the application going forward.