lark: fix db migrations in docker-compose env
fixes an error introduced 2398aad0. the database migration requires the
`postgres` user to have an the password set in `POSTGRES_ADMIN_PASSWORD`, but
our configuration tries to set it to *both* that and `POSTGRES_PASSWORD`. i
tried changing `POSTGRES_USER` but that didn't fix the error (i didn't try to
diagnose why). there's no reason the user in development can't just be the admin
all the time, so this will do the trick.

you can reproduce the error with:

`docker-compose down -v; docker-compose up`