lark-ui: setup docker-compose
borrow most of the docker setup from shoreline/starlight.

we use the multi-stage build, but drop some of the asset precompilation and, of
course, postgres (we'll go without a DB/ActiveRecord).

we need docker-compose, because `docker build .` isn't viable with
`$PROJECTPATH`. it's probably possible to use `docker build` with more
complicated arguments, but we're going to want other services sooner or later,
and this is more consistent with other projects.