surfliner: check authentication when using solr API for config
when we upload a solr configuration via the collections/configsets API, we want
to ensure the configset is "trusted". to ensure this, we need authentication
turned on.

the catch here is that the bitnimi solr chart doesn't turn on authentication on
an initial deploy until the all the nodes are up (it explictly waits for the
other nodes to become available). this means the nodes are briefly up with
authenication off(!) during deploy. assuming the init pod that uploads the
configset starts up quickly, this means it will upload to an unauthenticated
endpoint, and the configset will not be trusted.

to avoid this, we ensure authentication is on before uploading the configset.