shoreline: support custom config set names
with an external/preexisting Solr, it's necessary for the chart to allow custom
names for config sets. using the default `"solrconfig"` will cause multiple
deployments to step on each others' toes. having a custom name also helps in the
case that a preexisting configset and collection are in use.

actually, i think this last is necessary right now, since this chart won't
create the collection if its not present.

this could probably use some refinement, but i wanted to get it out there and
try to deploy into staging with this configuration. i created a staging
solrcloud on `dld-eks` to help with testing.