shoreline: use a custom PVC for staging geoserver at UCSB
migrate the old staging volume to a new `ebs-sc` volume.

created a snapshot of the older "in-tree" gp2 volume, then created a PVC to
provision one with `ebs-sc`, mounted both volumes to a temporary instance and
migrated the data from one to the other.

this isn't a system that will work at scale, but it should help us move
UCSB's shoreline prod volume in a similar way for a timely scale-up.

related to #723.