shoreline: add update strategy and priority class to Geoserver
adds new features to the geoserver chart.

allows update strategy to be set explictly to "Recreate". this is better for
this single pod  deployment. in cases where the volume won't freely attach to
multiple nodes, deployments can become stuck as the old pod hangs onto the
volume the new pod needs to successfully start up.

also adds a hook for passing in a `priorityClassName`, which can be useful to
manage scheduling policies.