comet: enable solr persistence in the development environment
turn on persistence for Solr in dev. it's attractive in development environments
to have persistence across deploys. this way developers can setup test data and
not need to recreate it repeatedly as they work.

k3d provides us local volume mappings as the default `StorageClass`, so just
turning on persistence for Solr adds persistence to the index.