comet: ensure the metadata db exists
connects to postgres and ensures the database for the metadata adapter exists.

we add add a db-wait container as well, to avoid this check failing if the DB
hasn't yet come up. maybe this is too much (it's a whole extra image and script
to support!) but it seems nice.