comet: add an AdministrativeSet seed
Hyrax demands an "AdministrativeSet" (`AdminSet` in ActiveFedora). These
determine the object workflow and policies in a broad way. We've talked about
calling these "Projects" in Comet. interestingly, i think i saw Northwestern
using similar terminology at the Samvera Partners meeting this week.

so here we go, an admin set. normally we wouldn't manually index (instead we'd
publish an event), but since this is a seed, we'll just shove this in the index
and avoid any other trappings.