comet: everyone can deposit
Hyrax has `everyone_can_deposit` but this is the heavier version.

the upstream implementation of this method checks that there is some
`registered_curation_concern` that the current ability `can?(:create,
MyModelClass)`, and then also checks that there's an `AdminSet` the user can
deposit to. this last step is costly: it involves a complex, difficult to
maintain, ActiveRecord query and then a secondary lookup in Solr to see if any
of any of the ids matched by AR identify AdminSets.

i don't think this pile of logic is going to be necessary for Comet. _if_ we
have users who can't deposit works, they'll probably be easily identified by
group/role (what is a group/role?): `student?`, or something.