comet: support LOCAL_VALUES_FILE in k3d deploy
It may sometimes be the case that the defaults specified in `k3d.yaml`
won't work for a particular local development environment use case. One
that's been raised is needing to change the ingress hostname.

This introduces support for a LOCAL_VALUES_FILE environment variable
that can act as an optional addition values file to pass to the `helm
deploy` command in `make deploy`