shoreline: allow opt-out of solr config set maintanance
adds flexibility to the helm chart by allowing users to skip loading the solr
configset to the solr instance. this is useful when using an existing external
solr is used and the configuration is to be maintained by a separate
process. especially: when it's not a Solr Cloud install, and there is no ZK.

connected to #711.