comet: build solr url in application; avoid `SOLR_URL`
the Hyrax chart has a helper for constructing the SOLR_URL. this seems like a
good idea at first blush, but it falls apart when the solr password is provided
by an external secret (and therefore not available in the chart).

we probably want a resolution for this on the Hyrax side. deprecating SOLR_URL
completely seems appropriate. in the meanwhile we should avoid its use here.