comet: ingest files to the storage backend
implement a custom extension of `Hyrax::WorkUploadsHandler` to attach filesets
and ingest files to storage inline. this avoids all the trouble of
`JobIoWrapper` and `IngestJob`. possibly we'll need to refine this for large
files, but the main reason for making this async in Hyrax is that
FileSet/ordered member attachment is very slow in `ActiveFedora`. we shouldn't
have that problem, so let's start simple.

`CometTransactionContainer` is used to inject the behavior to Hyrax`s
transaction registry. this is included in `initializers/hyrax` because trying to
put it in `/app` led to autoloader shenanigans. we can probably figure out a
better solution, but i wanted to unblock further work.