starlight: ensure migrations run before starlight test suite
Spotlight has an initializer that short-circuits/crashes when starting up
without a DB: https://github.com/projectblacklight/spotlight/issues/2133

when this happens, we fail to load critical code related to translations. in the
past, we've encountered this when running asset precompilation without a
database, e.g. in a docker build. now we're seeing this manifest for test
suites. the code never loads on startup, then the rails test schema management
code runs to get the migrations up to date. the test suite then fails because
critical mixin modules were never included.

for now, load the migration immeditately. we should seek to fix this
upstream. initializers being dependent on database state is... not great.