starlight: extract solr reindex to its own Job
The rest of the Starlight import process is about importing files and a
database backup. It was tightly coupled to solr reindexing via the rake
task that was called.

We should decouple this, so that folks running an external database, but
an in-chart deployment of Solr can make the choice to run reindexing
separately from database import.

We may want to further decouple these jobs, but this should get us past
an immediate issue that UCSD has with its infrastructure configuration.