starlight: remove search_spec system test
After discussion with Alexandra and Chrissy, we determined that this
test is not testing any functionality that is unique to Starlight. It is
only testing basic search functionality that is already tested upstream
in Spotlight.

Having said that, despite several attempts we were unable to get this
test passing in Spotlight 3. The offending expect() was trying to asset
that a custom metadata field (Identifier, identifier_tesim) was indexed
and searchable. This works perfectly fine in the UI, however the field
is not populated in the system test via CSV ingest.