comet: remove PATH update in Dockerfile
If we set the PATH to prefer the `DOCKERROOT/scripts` directory, then we
end up running the (same name) scripts from Surfliner for solrcloud
setup instead of the upstream scripts from Hyrax.

-ENV PATH="$DOCKERROOT/scripts:$PATH"

This `PATH` modification was only really useful in the test job, so to
address that we use the absolute path to those Surfliner scripts to
avoid any mistaken path issues should things shift location.