comet: move rspec examples.txt out of ./spec
When we run the tests in a k3d/k3s context, the current solution for
allowing the examples.txt file to be updated is chown'ing the entire
spec directory. The downside of this, is it creates a situation where
modifying any files on the host system can result in permission errors.

Instead, let's move that file out of the spec directory on the image
entirely, to the `./tmp` folder which is already chowned and gitignored.