comet: run our own db-setup initContainer in k3d
This is because there is currently an implicit ordering dependency in
our `ensure-metadata-db-exists` `extraInitContainer`. It assumes that it
can run _before_ the upstream `db-setup` `initContainer`, which in turn
runs ours local `db-migrate-seed.sh` script, which assumes that the
metadata database already exists.

To mitigate that, an upstream PR was introduced to allow downstream
Chart users, like us, to disable running the `db-setup` upstream
`initContainer` entirely.

https://github.com/samvera/hyrax/pull/5060