comet: pass ConfigMap and Secrets to db-setup
In our downstream `db-setup` `initContainer` we need access to a variety
of environment variables for all the initializers to be able to
successfully complete, allowing us to do the task of setting up the
databases.

This exposes everything the application Deployment container has access
to. It may be overkill, and we almost certainly want to find a way of
DRY'ing this out.