comet: don't define envFrom in db-setup
It turns out defining the `envFrom` array in the db-setup
`initContainer` and then trying to extend that array via set causes the
errors we're seeing.

```
spec.template.spec.initContainers[3].envFrom: Invalid value: "": may not
have more than one field specified at a time
```

What seems to work just fine is not defining the `envFrom` array in the
`initContainer` to begin with.