comet: manage hyrax chart versions in k3d
It can easily be the case where one might have an older version of the
Hyrax chart cached locally, and the version bump via `helm pull` does
not successfully get the new version in place.

This (simplistic) check against the cached `Chart.yaml` `version` should
hopefully be an improvement in this regard.