comet: upload the file first, then edit metadata
the file upload step takes more than a few ms. if the click-on-save is the very
next action, we're sometimes abusing the submit button in the tests.

if we upload the file first, then it can work while we're filling in other data
befor submitting the form. this seems to resolve race issues on this spec.