comet: use correct id for WorkflowRoles in seeds
`WorkflowResponsibility` wants a reference to a `WorkflowRole` (NOT a
`Role`). we were prevously passing `WorkflowRole#role_id` in as this
argument. this works ONLY in the case tha the `WorkflowRole` and `Role` objects
were created 1:1, sequentially, so their ids match.

that's not an assumption we can count on. in the case where it fails, we get
errors from ActiveRecord, or worse, just incorrect refernces to the wrong
roles.