comet: allow GenericObject to assign collection
We expect this to be addressed upstream, as what we've uncovered here is
that the GenericObjectForm does not actually pass the
`member_of_collection_ids` parameter to the controller when submitted.
Instead it populates the `member_of_collections_attributes` parameter
with the associated collection(s) identifiers and actions (delete, or
don't).

For now, we're changing our form to sync the populated
`member_of_collections_attributes` parameter to the
`member_of_collection_ids` paramter upon save/sync.