starlight: ucsb: bump puma max threads
we're seeing cases where the health checks appear to be failing due to
timeouts. it looks quite a bit like the reason for this is RIIIF eating up the
Puma threads stalled on longer running system calls for image manipulation.

on occassion, this leads to repeated failed health checks due to time out, and
consequent cycling of the pod, leading to some sporadic downtime. this doesn't
seem to come with any real resource encumberence for the pod, so it seems like a
good first step to just bump the thread count and see if that helps.

as follow up, we may want to consider:

  - introducing replicas for the rails pods (is this application stateless at
    this point?)
  - moving IIIF services into a separate container, even if that's still using
    RIIIF, and replicating those.