shoreline: add a hard override flag for disabling solr config init
when we moved to triggering this init container based on the Solr configuration,
we removed the ability to skip the config for cloud installations.

i think removing this tool is likely to bite us, since we won't have a way to
deploy without first configuring Solr. if we're having trouble with this
container or scripts, we'd have to shim this back in to get a healthy
deployment. let's just re-add it now.