comet: test having object in many collections
Ideally, we would setup the Collections in this spec using `let` and
`Hyrax.persister.save`. Unfortunately, due to the custom controller
handling we have for setting up the permissions for Collections, this
won't work without copying over all that code as well. So we're going
through the UI to set everything up and use our current custom
controller code in `#create`