WIP: comet: add unit tests for CollectionMembersController
we want to use this controller in place of the NestCollectionsController to
manage collections as members of collections. we also want to switch our local
version for the updated one upstream in Hyrax.

getting a test harness in and some clear tests for our local needs will help us
manage this controller as part of our app. that is: we feel like we own this
even though it's provided by the Hyrax Engine.