comet: extract pod review limit defaults
UCSD was running into an issue of hitting up against these limits in
review apps, which seemed to cause a SIGTERM thrown against the
container running in the pod that hit this limit.

Make these limits optional, including for UCSB since it doesn't seem to
be an issue on that end.