comet: don't create STAGING_AREA_S3_BUCKET
This code presumes a running Minio instance. In an initializer, this is
probably not a great idea, unless we want to loop and wait until we're
sure there's a connection.

The pattern we've used elsewhere is to either:
- have an initContainer or separate task responsible for setting up
  thigns like buckets.
- assume the deployer is responsible for ensuring buckets required for
  an application already exist

For docker compose and review apps we have a solution for this already
in place