starlight: add ability to force backup loads on existing systems
the procedure default procedure for loading backups only runs as a helm
post-install hook. this seems good in that it avoids accidental overrides of
production data.

however, it's a common use case to *want* to load a backup into a running
system. for this case, this provides a `starlight.backups.force` flag which will
run as a post-upgrade hook.

adding the `before-hook-creation` delete policy allows a new deployment to retry
loading from backup in the case that a previous hook failed.