starlight: inline solr reindex
the reindex task included in Spotlight enqueues a series of background
jobs (one for each exhibit) and reports success regardless of whether those jobs
succeed. Spotlight provides tools for tracking progress, of those jobs, which is
pretty cool!

but for a standalone K8S job, all this just layers aynchronous processing over
more asynchronous processing, and it makes it hard to notice what's happening
with the task you're trying to run. running inline in this context is nicer, so
we provide a job for that here.