surfliner: tag images with git tags if used
when we do a tagged build, tag the built image with the given tag.

e.g. when we push a tag `starlight-v0.1.1` we want to end up with an image
tagged with something associated with the tag (rather than the commit sha).

this just reuses the tag name resulting in an image like
`surfliner/starlight:starlight-v0.1.1`. we could consider doing some string
manipulation, but i thought it better to separate that from the initial proposal
to tag these images.