starlight: scale worker pods separately from webapp pods
the factors that cause workers and webapp pods to need to scale are mostly
unrelated to each other. make them scale on separate variables.

to retain compatibility for existing setups, let the `replicaCount` serve as a
default. we probably want to recommend folks use a `workerReplicaCount`
explictly, unless their `replicaCount` is 1.