starlight: revert back to Spotlight reindex task
Our local `reindex_now` rake task appears to cause issues with
reindexing Spotlight 3.x content, at least some of the content in the
existing UCSD deployment.

We revert back to using the upstream rake task until we have more time
to address this and get the inline indexing we want.