comet: refactor rabbitmq work for tidewater to use Dry::Events
per discussion with @kibigo and @lsitu, refactoring the MQ events behind an
in-app Dry::Events bus. Hyrax already has this bus in
place (https://github.com/samvera/hyrax/wiki/Hyrax's-Event-Bus-(Hyrax::Publisher)),
and reusing it as the seam for events going into RabbitMQ should be very
flexible.

our hope is that developers will only need to worry about RabbitMQ
implementation details when working on relevant Listener implementations. in the
rest of the codebase they should be able to rely on publishing to the
`Hyrax.publisher` with a standardized payload, and go about their day.

it's approachable to subscribe/unsubscribe Dry::Events at runtime, so unit tests
can box off the message queue.

---

only run tests that interact with the queue from the system specs. this avoids
the need to ship a RabbitMQ when running unit tests.