tidewater: rename OAI_REPOSITORY_URL
As noted in #944 our defined application route did not align with what
we were setting as a default in our `.env`.

We are extracting the variable portion of the URL to a new environment
variable named `OAI_REPOSITORY_ORIGIN` and fixing the `oai` path. We
also add some tests to ensure this routing is set and matches what is
specified in the provider configuration.