starlight:ucsb: set staging values for UCSB Starlight Stage
use a tagged image.

get a few starlight staging replicas running to handle traffic spikes from IIIF
a bit better.

drop the `starlight-staging.eks` domain, which doesn't seem to be able to serve
out images (what's the reason for this?).

add resource requests and a priority class for memcached.

remove settings for the data migration from the legacy VM.