comet: mount current directory to webapp, not engine
in some of our support containers, we were mounting the current app to
`hyrax-engine` (which is an unused directory), instead of `hyrax-webapp`. this
meant that the current codebase wasn't getting used for these support tasks,
only whatever was present at build time for the container image.

mount the current dev path to the correct location in the image for a better dev
experience!