comet: put collection publish behavior behind a feature flag
Collection publish behavior needs some more consideration to be ready to merge
and deploy to prod.

for example, at present:
  - ANYONE can publish
  - we don't save any data about the collection's published/unpublished status,
  nor that of any of their objects.

we want to get this merged to support further development on the RabbitMQ
message broker, and so we can demo it for discussion. but until it's more
mature, we doubt it will make it past staging.