comet: don't depend on `rabbitmq` for `docker-compose` db migrations
but DO depend on it for the other containers, which hang during initialization
if `RABBITMQ_ENABLED=true` but `rabbitmq` isn't started.

we may need to follow up to make the initializer more tolerant. we may also want
to disable RabbitMQ in the migrations more generally (via a chart variable? or
maybe there's a way we can do this with `rake`'s environment?)